pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "hvym_stellar"
authors = [{name = "Fibo Metavinci", email = "pszdw-75nat-5227i-bha5s-y7lai-pebdq-o2agp-3xho4-hd6z6-emxrd-nqe@dmail.ai"}]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
dynamic = ["version", "description"]
dependencies = [
    "annotated-types>=0.7.0",
    "biscuit-python",
    "certifi>=2025.4.26",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.2",
    "cryptography>=46.0.3",
    "docutils>=0.21.2",
    "idna>=3.10",
    "mnemonic>=0.20",
    "pycparser>=2.22",
    "pydantic>=2.11.5",
    "pydantic_core>=2.33.2",
    "pymacaroons>=0.13.0",
    "PyNaCl>=1.5.0",
    "pytest>=9.0.2",
    "requests>=2.32.4",
    "requests-sse>=0.5.1",
    "six>=1.17.0",
    "stellar-sdk>=12.3.0",
    "toml>=0.10.2",
    "tomli_w>=1.2.0",
    "typing-inspection>=0.4.1",
    "typing_extensions>=4.14.0",
    "urllib3>=2.5.0",
    "xdrlib3>=0.1.1",
]

[project.urls]
Home = "https://github.com/inviti8/hvym_stellar"

[dependency-groups]
dev = []
